+++ src/index.ts
@@ -136,3 +136,6 @@ export interface HttpClient {
 }
 
 export { makeHttpClient } from './core/http';
+
+export { LocalStorage, SessionStorage } from './core/storage';
+export { AsyncFileReader } from './core/file';