+++ CHANGELOG.md
@@ -1,19 +1,23 @@
 # CHANGELOG
 
 ## V0.2.1
+### Changed
+- Storage: *get* replaces *getString* and *getObject*
+
 ### Refactored
 - HTTP: Check the enviroment NODE_ENV and MOCK_DISABLED
 
 ### Removed
-- HTTP: enableMock
+- HTTP: *enableMock*
+- Storage: *getString* and *getObject*
 
 ## V0.2.0 ｜ 2020-11-16
 ### Features
-- HTTP: API's mock, enableMock
 - AsyncFileReader
+- HTTP: API's mock, *enableMock*
 
 ### Changed
-- HTTP: The parameter request in response and error of API.
+- HTTP: The parameter *request* in *response* and *error* of API.
 
 ### Removed
 - Model: put, loading